refactor(migrations): address PR review on 002 canonical-port discovery
- Reuse extractPgservePortFromStatus (moved to src/lib/pgserve-status.ts,
  re-exported from update.ts) so the migration tolerates nested
  instance.port / runtime.port status shapes — reading only top-level .port
  would null-resolve and make the migration a permanent no-op on those hosts
  (Codex P1).
- Guard the parsed port with > 0 so Number(null|false|'')===0 is rejected
  (Gemini).
- selectLegacyEmbedded now returns ALL non-canonical postmasters (filter, not
  find); apply stops each and validate waits for all to exit — otherwise a
  second stray survives and validate fails after 5s (Gemini).
